requirements.txt
pandas==1.1.0
python-dotenv==0.14.0
schedule==0.6.0
selenium==3.141.0
tqdm==4.48.2
